updates for duplicate veto
git-svn-id: svn+ssh://svn.cern.ch/reps/atlasphys-hsg8/Physics/Higgs/HSG8/AnalysisCode/multileptons/xAOD/ttHMultiAna/trunk@236761 883ba7d9-fdd0-4202-9341-49aa55999ad8

+++ ttHMultilepton/DuplicateSelector.h
@@ -6,27 +6,17 @@
 #include <set>
 
 /**
- * @brief Create HWW style Count histogram
+ * @brief keep track of dupes
  */
 class DuplicateSelector : public top::EventSelectorBase {
 public:
-    /**
-     * @brief Setup some example plots.
-     *
-     * Add a bunch of histograms.
-     *
-     * @param name The name of the directory to store histograms in, in the
-     * output file.  e.g. you might have ee, mumu and emu.
-     * @param outputFile The output file. Needs setting up at the very start
-     * so that we can attach the files.
-     * @param wk Only used by EventLoop, ok as nullptr as default.
-     */
+
   DuplicateSelector();
 
     /**
-     * @brief Fill the histograms.
+     * @brief Check dupes
      *
-     * @return True because it doesn't select any events.
+     * @return True if event is not dupe
      */
     virtual bool apply(const top::Event& event) const override;
 
@@ -40,7 +30,7 @@ public:
 private:
 
     // all processed events
-    mutable std::set<std::pair<int, int> > m_processedEvents;
+    mutable std::set<std::pair<int, uint64_t> > m_processedEvents;
   
     // Nominal hash value
     std::size_t m_nominalHashValue;
